Fix QPlatformTheme::IconThemeSearchPaths test
Previously, the test did toString().isEmpty, which would usually succeed
simply because a multiple-element stringlist casts to an empty string.
However, it would fail if the list was one element (such as if
XDG_DATA_DIRS were set to a single directory, and ~/.icons did not
exist).

Either way, it was not testing what it should.  This actually tests the
returned value to some extent.

Reviewed by: Alex Fiestas