Pass the KDE_COLOR_SCHEME_PATH application property to toplevel windows
This feature is upstreamed from the Oxygen style. It is a general
feature for integration with our window manager and thus should be in
the generic style implementation and not just in Oxygen.

So far the implementation only supports X11 and installs an X property
on all toplevel windows. On other platforms it is currently a no-op.

This introduces an optional dependency on XCB and QtX11Extras. If they
are not found the feature is obviously not compiled.

REVIEW: 115285