Use a QPointer for QMenu of SystemTrayMenu
We need better memory management as the QMenu the SystemTrayMenu creates
might be deleted externally causing a double delete if the so-far
QScopedPoiter cleans up.

Switch to QPointer and call deleteLater if the pointer is still valid
during destroying the SystemTrayMenu.

BUG: 343976
REVIEW: 122978
CHANGELOG: Fix possible crash when destroying a QSystemTrayIcon (triggered by Trojita)