Fix kdeplatformtheme_unittest after last commit
Use Q_CONSTRUCTOR_FUNCTION for prepareEnvironment, so the QApplication's
platform theme is using the test paths.

This also fixes updating the WheelScrollLines when the settings change,
because QApplication is only using theme hint on startup. If the value changes,
we need to call QApplication::setWheelScrollLines instead of just updating
the theme hint.

REVIEW: 125877