Add "applicationcrash" notification event
This is going to be used by DrKonqi. The "Catastrophe" event, while mentioning
"something bad happened, at least the application crashed", draws a lot of attention
to it with its sound. Also, having a dedicated notification for this enables users
to specifically disable DrKonqi messages in case they really don't care.

REVIEW: 126566