Remove platformtheme plugin
The platformthem plugin got split out of frameworksintegration and
lives in plasma-integration now. As it was for Plasma and has
dependencies on Plasma (e.g. Breeze) it should be removed from
frameworksintegration.

Plasma-integration had a release in the mean-time.
Note for distributions: Don't combine this frameworksintegration with
Plasma < 5.6!

CHANGELOG: Remove platformtheme plugin as it's in plasma-integration

REVIEW: 127756