plasma_workspace.notifyrc: relabel to match button
The HIG recommend that buttons labeled with locations match the title of
the page they open. In plasma/plasma-workspace!4000, a specific button
was introduced to open the Plasma Workspace notification settings, using
the "System Notifications" label as a more user-centered term that puts
importance on the things the user can configure, rather than the
implementation detail that these settings are part of the Plasma
Workspace notifications. This, however, has led to the button label not
matching the title of the page it links to.

This change modifies plasma_workspace.notifyrc to also use the "System
Notifications" terminology.