Turn on popups for device notifications
Part of https://invent.kde.org/plasma/plasma-desktop/-/issues/149

These notifications are a useful way to communicate to the user that
their device isn't dead and was actuallt detected. Currently we have
only sounds turned on, and no popups. This problematic for
accessibility because deaf people won't hear the sound.

For this reason, EU Directive 2019/882 requires that all sound-based
notifications be accompanied by something visual as well.

Let's turn on popups to comply with that. We did some work to make the
popups not annoying, including:
- Not appearing in the history
- Removing the connection notification when disconnected (and vice
  versa)
- Removing the notification when it's for a USB disk and the Disks &
  Devices widget appears automatically.